Move SPI implementations to provider package
