+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/java/io/opentelemetry/sdk/autoconfigure/provider/FirstCustomizerProvider.java
@@ -3,23 +3,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure;
+package io.opentelemetry.sdk.autoconfigure.provider;
 
 import static java.util.Collections.singletonMap;
 
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizer;
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizerProvider;
 
-public class ThirdCustomizerProvider implements AutoConfigurationCustomizerProvider {
+public class FirstCustomizerProvider implements AutoConfigurationCustomizerProvider {
 
   @Override
   public void customize(AutoConfigurationCustomizer autoConfiguration) {
-    autoConfiguration.addPropertiesSupplier(
-        () -> singletonMap("otel.from-config", "configured value"));
+    autoConfiguration.addPropertiesSupplier(() -> singletonMap("otel.from-config", "unused"));
   }
 
   @Override
   public int order() {
-    return 3000;
+    return 1000;
   }
 }
