+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/java/io/opentelemetry/sdk/autoconfigure/provider/FirstResourceProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure;
+package io.opentelemetry.sdk.autoconfigure.provider;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 
@@ -12,15 +12,20 @@ import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider;
 import io.opentelemetry.sdk.resources.Resource;
 
-public class ThirdResourceProvider implements ResourceProvider {
+public class FirstResourceProvider implements ResourceProvider {
 
   @Override
   public Resource createResource(ConfigProperties config) {
-    return Resource.create(Attributes.of(stringKey("otel.some_resource"), "real value"));
+    return Resource.create(
+        Attributes.of(
+            stringKey("otel.some_resource"),
+            "unused",
+            stringKey("otel.from_config"),
+            config.getString("otel.from-config", "default")));
   }
 
   @Override
   public int order() {
-    return 300;
+    return 100;
   }
 }
