+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/java/io/opentelemetry/sdk/autoconfigure/provider/SecondCustomizerProvider.java
@@ -3,22 +3,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure;
+package io.opentelemetry.sdk.autoconfigure.provider;
 
 import static java.util.Collections.singletonMap;
 
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizer;
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizerProvider;
 
-public class FirstCustomizerProvider implements AutoConfigurationCustomizerProvider {
+public class SecondCustomizerProvider implements AutoConfigurationCustomizerProvider {
 
   @Override
   public void customize(AutoConfigurationCustomizer autoConfiguration) {
-    autoConfiguration.addPropertiesSupplier(() -> singletonMap("otel.from-config", "unused"));
+    autoConfiguration.addPropertiesSupplier(() -> singletonMap("otel.from-config", "still unused"));
   }
 
   @Override
   public int order() {
-    return 1000;
+    return 2000;
   }
 }
