+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/java/io/opentelemetry/sdk/autoconfigure/provider/SecondResourceProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure;
+package io.opentelemetry.sdk.autoconfigure.provider;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 
@@ -12,20 +12,15 @@ import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider;
 import io.opentelemetry.sdk.resources.Resource;
 
-public class FirstResourceProvider implements ResourceProvider {
+public class SecondResourceProvider implements ResourceProvider {
 
   @Override
   public Resource createResource(ConfigProperties config) {
-    return Resource.create(
-        Attributes.of(
-            stringKey("otel.some_resource"),
-            "unused",
-            stringKey("otel.from_config"),
-            config.getString("otel.from-config", "default")));
+    return Resource.create(Attributes.of(stringKey("otel.some_resource"), "still unused"));
   }
 
   @Override
   public int order() {
-    return 100;
+    return 200;
   }
 }
