+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/java/io/opentelemetry/sdk/autoconfigure/provider/ThirdCustomizerProvider.java
@@ -3,22 +3,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.autoconfigure;
+package io.opentelemetry.sdk.autoconfigure.provider;
 
 import static java.util.Collections.singletonMap;
 
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizer;
 import io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizerProvider;
 
-public class SecondCustomizerProvider implements AutoConfigurationCustomizerProvider {
+public class ThirdCustomizerProvider implements AutoConfigurationCustomizerProvider {
 
   @Override
   public void customize(AutoConfigurationCustomizer autoConfiguration) {
-    autoConfiguration.addPropertiesSupplier(() -> singletonMap("otel.from-config", "still unused"));
+    autoConfiguration.addPropertiesSupplier(
+        () -> singletonMap("otel.from-config", "configured value"));
   }
 
   @Override
   public int order() {
-    return 2000;
+    return 3000;
   }
 }
