+++ sdk-extensions/autoconfigure/src/testAutoConfigureOrder/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider
@@ -1,3 +1,3 @@
-io.opentelemetry.sdk.autoconfigure.ThirdResourceProvider
-io.opentelemetry.sdk.autoconfigure.SecondResourceProvider
-io.opentelemetry.sdk.autoconfigure.FirstResourceProvider
+io.opentelemetry.sdk.autoconfigure.provider.ThirdResourceProvider
+io.opentelemetry.sdk.autoconfigure.provider.SecondResourceProvider
+io.opentelemetry.sdk.autoconfigure.provider.FirstResourceProvider
