+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizerProvider
@@ -1,3 +1,3 @@
-io.opentelemetry.sdk.autoconfigure.SpanExporterCustomizer
-io.opentelemetry.sdk.autoconfigure.MetricCustomizer
-io.opentelemetry.sdk.autoconfigure.LogRecordExporterCustomizer
+io.opentelemetry.sdk.autoconfigure.provider.SpanExporterCustomizer
+io.opentelemetry.sdk.autoconfigure.provider.MetricCustomizer
+io.opentelemetry.sdk.autoconfigure.provider.LogRecordExporterCustomizer
