+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider
@@ -1,2 +1,2 @@
-io.opentelemetry.sdk.autoconfigure.TestAnimalResourceProvider
-io.opentelemetry.sdk.autoconfigure.TestColorResourceProvider
+io.opentelemetry.sdk.autoconfigure.provider.TestAnimalResourceProvider
+io.opentelemetry.sdk.autoconfigure.provider.TestColorResourceProvider
