+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.logs.ConfigurableLogRecordExporterProvider
@@ -1,2 +1,2 @@
-io.opentelemetry.sdk.autoconfigure.TestConfigurableLogRecordExporterProvider
-io.opentelemetry.sdk.autoconfigure.ThrowingConfigurableLogRecordExporterProvider
+io.opentelemetry.sdk.autoconfigure.provider.TestConfigurableLogRecordExporterProvider
+io.opentelemetry.sdk.autoconfigure.provider.ThrowingConfigurableLogRecordExporterProvider
