+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.metrics.ConfigurableMetricExporterProvider
@@ -1,2 +1,2 @@
-io.opentelemetry.sdk.autoconfigure.TestConfigurableMetricExporterProvider
-io.opentelemetry.sdk.autoconfigure.ThrowingConfigurableMetricExporterProvider
+io.opentelemetry.sdk.autoconfigure.provider.TestConfigurableMetricExporterProvider
+io.opentelemetry.sdk.autoconfigure.provider.ThrowingConfigurableMetricExporterProvider
