+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.traces.ConfigurableSamplerProvider
@@ -1,2 +1,2 @@
-io.opentelemetry.sdk.autoconfigure.TestConfigurableSamplerProvider
-io.opentelemetry.sdk.autoconfigure.ThrowingConfigurableSamplerProvider
+io.opentelemetry.sdk.autoconfigure.provider.TestConfigurableSamplerProvider
+io.opentelemetry.sdk.autoconfigure.provider.ThrowingConfigurableSamplerProvider
