+++ sdk-extensions/autoconfigure/src/testFullConfig/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.traces.SdkTracerProviderConfigurer
@@ -1,1 +1,1 @@
-io.opentelemetry.sdk.autoconfigure.TestTracerProviderConfigurer
+io.opentelemetry.sdk.autoconfigure.provider.TestTracerProviderConfigurer
